+++ src/services/continuous_quality_monitor/__init__.py
@@ -1,0 +1,5 @@
+"""Continuous quality monitor package."""
+
+from .monitor import ContinuousQualityMonitor
+
+__all__ = ["ContinuousQualityMonitor"]
